requirements.txt
Flask==3.0.3
openai==1.42.0
google-cloud-vision==3.7.4
python-dotenv==1.0.1
